Moved to Apply Damage Model
