+++ CSharpSourceCode/Extensions/AgentExtensions.cs
@@ -170,308 +170,21 @@ namespace TOR_Core.Extensions
         /// Unit information gets read out from the tor_extendedunitproperties.xml
         /// </param>
         /// <paramref name="agent"/>
-        /// <param name="propertyMask">Which properties needs to be accessed? be mindful about which information is really required for your specific situation. Not used properties are returned as empty arrays</param>
+        /// <param name="propertyMask">Which properties needs to be accessed? Be mindful about which information is really required for your specific situation. Not used properties are returned as empty arrays</param>
         ///<returns>A struct containing 3 arrays, each containing the  properties:  proportions, damage amplifications and resistances.</returns>
         public static AgentPropertyContainer GetProperties(this Agent agent, PropertyMask propertyMask, AttackTypeMask attackTypeMask)
         {
-            if (agent.IsMount)
+            if (!(MissionGameModels.Current.AgentApplyDamageModel is TORAgentApplyDamageModel damageModel))
+            {
                 return new AgentPropertyContainer();
-
-            float[] damageProportions = new float[(int)DamageType.All + 1];
-            float[] damageAmplifications = new float[(int)DamageType.All + 1];
-            float[] damageResistances = new float[(int)DamageType.All + 1];
-            float[] additionalDamagePercentages = new float[(int)DamageType.All + 1];
+            }
             
+            return damageModel.CreateAgentPropertyContainer(agent, propertyMask, attackTypeMask);
             
-            if (!agent.IsHero)
-            {
-                AssignUnitProperties();
-            }
-
-            else
-            {
-                AssignCharacterProperties();
-            }
-
-            var result = new AgentPropertyContainer(damageProportions, damageAmplifications, damageResistances, additionalDamagePercentages);
-
-            if (Game.Current.GameType is Campaign)
-            {
-                var model = MissionGameModels.Current.AgentStatCalculateModel as TORAgentStatCalculateModel;
-                if(model != null)
-                {
-                    result = model.AddPerkEffectsToAgentPropertyContainer(agent, propertyMask, attackTypeMask, result);
-                }
-            }
-
-            return result;
-            
-            void AssignUnitProperties()
-            {
-                 if (propertyMask == PropertyMask.Attack || propertyMask == PropertyMask.All)
-                 {
-                     var unitDamageProportion = agent.Character.GetUnitDamageProportions();
-                     foreach (var proportionTuple in unitDamageProportion)
-                     {
-                         damageProportions[(int)proportionTuple.DamageType] = proportionTuple.Percent;
-                     }
-                     var offenseProperties = agent.Character.GetAttackProperties();
-
-                     //add all offense properties of the Unit
-                     foreach (var property in offenseProperties)
-                     {
-                         damageAmplifications[(int)property.AmplifiedDamageType] += property.DamageAmplifier;
-                     }
-                     //add temporary effects like buffs to attack bonuses on items
-                     List<ItemTrait> dynamicTraits = agent.GetComponent<ItemTraitAgentComponent>()
-                         .GetDynamicTraits(agent.WieldedWeapon.Item);
-                     foreach (var dynamicTrait in dynamicTraits)
-                     {
-                         var attackProperty = dynamicTrait.AmplifierTuple;
-
-                         if (attackProperty != null)
-                         {
-                             damageAmplifications[(int)attackProperty.AmplifiedDamageType] += attackProperty.DamageAmplifier;
-                         }
-                         var additionalDamageProperty = dynamicTrait.AdditionalDamageTuple;
-                         if (additionalDamageProperty != null)
-                         {
-                             additionalDamagePercentages[(int)additionalDamageProperty.DamageType] += additionalDamageProperty.Percent;
-                         }
-                     }
-                     var statusEffectAmplifiers = agent.GetComponent<StatusEffectComponent>().GetAmplifiers(attackTypeMask);
-                     for (int i = 0; i < damageAmplifications.Length; i++)
-                     {
-                         damageAmplifications[i] += statusEffectAmplifiers[i];
-                     }
-                 }
-                 if (propertyMask == PropertyMask.Defense || propertyMask == PropertyMask.All)
-                 {
-                     //add all defense properties of the Unit
-                     var defenseProperties = agent.Character.GetDefenseProperties();
-
-                     foreach (var property in defenseProperties)
-                     {
-                         damageResistances[(int)property.ResistedDamageType] += property.ReductionPercent;
-                     }
-
-                     //add temporary effects like buffs to defense bonuses
-                     List<ItemTrait> dynamicTraits = agent.GetComponent<ItemTraitAgentComponent>()
-                         .GetDynamicTraits(agent.WieldedWeapon.Item);
-
-                     foreach (var dynamicTrait in dynamicTraits)
-                     {
-                         var defenseProperty = dynamicTrait.ResistanceTuple;
-                         if (defenseProperty != null)
-                         {
-                             damageResistances[(int)defenseProperty.ResistedDamageType] += defenseProperty.ReductionPercent;
-                         }
-                     }
-
-                     //status effects
-                     var statusEffectResistances = agent.GetComponent<StatusEffectComponent>().GetResistances(attackTypeMask);
-
-                     for (int i = 0; i < damageResistances.Length; i++)
-                     {
-                         damageResistances[i] += statusEffectResistances[i];
-                     }
-                 }
-            }
-
-            void AssignCharacterProperties()
-            {
-                if (propertyMask == PropertyMask.Attack || propertyMask == PropertyMask.All)
-                {
-                    //Hero item level attributes 
-                    List<ItemTrait> itemTraits = new List<ItemTrait>();
-                    List<ItemObject> armorItems;
-                    // get all equipment Pieces - here only armor
-                    armorItems = agent.Character.GetCharacterEquipment(EquipmentIndex.ArmorItemBeginSlot);
-                    foreach (var item in armorItems)
-                    {
-                        if (item.HasTrait())
-                            itemTraits.AddRange(item.GetTraits(agent));
-                    }
-                    //equipment amplifiers, also implies dynamic traits
-                    foreach (var itemTrait in itemTraits)
-                    {
-                        var property = itemTrait.AmplifierTuple;
-                        if (property != null)
-                            damageAmplifications[(int)property.AmplifiedDamageType] += property.DamageAmplifier;
-
-                        var additionalDamageProperty = itemTrait.AdditionalDamageTuple;
-                        if (additionalDamageProperty != null)
-                        {
-                            additionalDamagePercentages[(int)additionalDamageProperty.DamageType] += additionalDamageProperty.Percent;
-                        }
-
-                    }
-
-                    var statusEffectAmplifiers = agent.GetComponent<StatusEffectComponent>().GetAmplifiers(attackTypeMask);
-
-                    for (int i = 0; i < damageAmplifications.Length; i++)
-                    {
-                        damageAmplifications[i] += statusEffectAmplifiers[i];
-                    }
-                    
-                    //Weapon properties
-                    if (attackTypeMask == AttackTypeMask.Ranged)
-                    {
-                        if (agent.WieldedWeapon.Item != null)
-                        {
-                            var ammoItem = Mission.Current.Missiles.FirstOrDefault(x => x.ShooterAgent == agent)?.Weapon.Item;
-                            var weapon = agent.WieldedWeapon.Item;
-                            List<ItemTrait> rangeItemTraits = new List<ItemTrait>();
-                            
-                            if(ammoItem!=null)
-                                rangeItemTraits.AddRange(ammoItem.GetTraits());
-                            rangeItemTraits.AddRange(weapon.GetTraits());
-                            foreach (var itemTrait in rangeItemTraits)
-                            {
-                                var property = itemTrait.AmplifierTuple;
-                                if(property!=null)
-                                    damageAmplifications[(int)property.AmplifiedDamageType] += property.DamageAmplifier;
-
-                                var additionalDamageProperty = itemTrait.AdditionalDamageTuple;
-                                if (additionalDamageProperty != null)
-                                {
-                                    additionalDamagePercentages[(int)additionalDamageProperty.DamageType] += additionalDamageProperty.Percent;
-                                }
-                            }
-                            //range damage Propotions
-                            var weaponProperty = weapon.GetTorSpecificData().DamageProportions;
-                            if (weaponProperty != null)
-                            {
-                                foreach (var tuple in weaponProperty)
-                                {
-                                    damageProportions[(int)tuple.DamageType] = tuple.Percent;
-                                }
-                            }
-                        }
-                        else
-                        {
-                            damageProportions[(int)DamageType.Physical] = 1f; //memo , this is for siege weapons
-                        }
-                    }
-
-                    if (attackTypeMask == AttackTypeMask.Melee)
-                    {
-                        if (agent.WieldedWeapon.Item != null)
-                        {
-                            var weapon = agent.WieldedWeapon.Item;
-                            var offhand = agent.WieldedOffhandWeapon.Item;
-                            List<ItemTrait> meleeItemTraits = new List<ItemTrait>();
-                            meleeItemTraits.AddRange(weapon.GetTraits());
-                            if (offhand != null)
-                                meleeItemTraits.AddRange(offhand.GetTraits());
-                            
-                            foreach (var itemTrait in meleeItemTraits)
-                            {
-                                var property = itemTrait.AmplifierTuple;
-                                
-                                if(property!=null)
-                                    damageAmplifications[(int)property.AmplifiedDamageType] += property.DamageAmplifier;
-
-                                var additionalDamageProperty = itemTrait.AdditionalDamageTuple;
-                                if (additionalDamageProperty != null)
-                                {
-                                    additionalDamagePercentages[(int)additionalDamageProperty.DamageType] += additionalDamageProperty.Percent;
-                                }
-                            }
-                            
-                            var weaponProperty = weapon.GetTorSpecificData().DamageProportions;
-                            if (weaponProperty != null)
-                            {
-                                foreach (var tuple in weaponProperty)
-                                {
-                                    damageProportions[(int)tuple.DamageType] = tuple.Percent;
-                                }
-                            }
-                        }
-                        else
-                        {
-                            damageProportions[(int)DamageType.Physical] = 1f; //memo , this is for siege weapons, in principle a wielded Item shouldn't be found either in case of spell casting - yet it is found.
-                        }
-                    }
-                    
-                    if (attackTypeMask == AttackTypeMask.Spell)
-                    {
-                        if (!agent.WieldedOffhandWeapon.IsEmpty && agent.WieldedOffhandWeapon.Item != null && agent.WieldedOffhandWeapon.Item.StringId.Contains("staff"))
-                        {
-                            List<ItemTrait> staffItemTraits = agent.WieldedOffhandWeapon.Item.GetTraits();
-
-                            foreach (var itemTrait in staffItemTraits)
-                            {
-                                var property = itemTrait.AmplifierTuple;
-                                
-                                if(property!=null)
-                                    damageAmplifications[(int)property.AmplifiedDamageType] += property.DamageAmplifier;
-
-                                var additionalDamageProperty = itemTrait.AdditionalDamageTuple;
-                                if (additionalDamageProperty != null)
-                                {
-                                    additionalDamagePercentages[(int)additionalDamageProperty.DamageType] += additionalDamageProperty.Percent;
-                                }
-                            }
-                        }
-                    }
-                    
-                }
-                if (propertyMask == PropertyMask.Defense || propertyMask == PropertyMask.All)
-                {
-                    //Hero item level attributes 
-
-                    List<ItemTrait> itemTraits = new List<ItemTrait>();
-                    List<ItemObject> items;
-
-                    items = agent.Character.GetCharacterEquipment(EquipmentIndex.ArmorItemBeginSlot);
-                    foreach (var item in items)
-                    {
-                        if (item.HasTrait())
-                            itemTraits.AddRange(item.GetTraits(agent));
-                    }
-
-                    //equipment resistances , also implies dynamic traits
-                    foreach (var itemTrait in itemTraits)
-                    {
-                        var defenseProperty = itemTrait.ResistanceTuple;
-                        if (defenseProperty == null)
-                            continue;
-                        damageResistances[(int)defenseProperty.ResistedDamageType] += defenseProperty.ReductionPercent;
-                    }
-
-                    //statuseffects
-                    var statusEffectResistances = agent.GetComponent<StatusEffectComponent>().GetResistances(attackTypeMask);
-
-                    for (int i = 0; i < damageResistances.Length; i++)
-                    {
-                        damageResistances[i] += statusEffectResistances[i];
-                    }
-
-                    if (agent.WieldedWeapon.Item != null)
-                    {
-                        List<ItemTrait> wieldedItemTraits = new List<ItemTrait>();
-                        var weapon = agent.WieldedWeapon.Item;
-
-                        var offHand = agent.WieldedOffhandWeapon.Item;
-                        
-                        wieldedItemTraits.AddRange(weapon.GetTraits());
-                        if (offHand != null)
-                        {
-                            wieldedItemTraits.AddRange(offHand.GetTraits());
-                        }
-                        
-                        foreach (var itemTrait in wieldedItemTraits)
-                        {
-                            var defenseProperty = itemTrait.ResistanceTuple;
-                            if (defenseProperty == null)
-                                continue;
-                            damageResistances[(int)defenseProperty.ResistedDamageType] += defenseProperty.ReductionPercent;
-                        }   
-                    }
-                }
-            }
         }
+        
+        
+        
 
         public static Ability GetCurrentAbility(this Agent agent)
         {
