+++ CSharpSourceCode/HarmonyPatches/DamagePatch.cs
@@ -40,13 +40,8 @@ namespace TOR_Core.HarmonyPatches
             AttackTypeMask attackTypeMask = DetermineMask(b);
 
             float[] damageCategories = new float[(int)DamageType.All + 1];
-            var attackerPropertyContainer = AgentPropertyContainer.InitNew();
-            var victimPropertyContainer = AgentPropertyContainer.InitNew();
-            if (!Mission.Current.IsArenaMission())
-            { 
-                attackerPropertyContainer = attacker.GetProperties(PropertyMask.Attack, attackTypeMask); 
-                victimPropertyContainer = victim.GetProperties(PropertyMask.Defense, attackTypeMask);
-            }
+            var attackerPropertyContainer = attacker.GetProperties(PropertyMask.Attack, attackTypeMask); 
+            var victimPropertyContainer = victim.GetProperties(PropertyMask.Defense, attackTypeMask);
             
             var friendlyFire = attacker.Team == victim.Team;
             var damageProportions = new float[7];
